+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add game addon remote client compatibility check (to check if client has the DLC if the server has it)
 - Add player skin support
 - Add player exp info to save file and remote syncing
+- Added the party system
 
 ### Fixed